Widgets // Unify widget background API.
